# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API Teste Helpper",
    "description": "Documentation automatically generated by the <b>swagger.autogen</b> module."
  },
  "host": "localhost:44340",
  "basePath": "/api",
  "tags": [],
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/signin": {
      "post": {
        "tags": [
          "Authorization"
        ],
        "description": "Endpoint para login.",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "Informações do usuário.",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/SignIn"
            }
          }
        ],
        "responses": {}
      }
    },
    "/users": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint para obter um usuário.",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "api_key": {
              "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
              }
            }
          }
        ]
      }
    },
    "/users/{userAdd}": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint para obter um usuário.",
        "parameters": [
          {
            "name": "userAdd",
            "in": "body",
            "required": true,
            "description": "Informações do usuário.",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/AddUser"
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "api_key": {
              "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
              }
            }
          }
        ]
      }
    },
    "/users/{userUpdate}": {
      "put": {
        "tags": [
          "User"
        ],
        "description": "Endpoint para alterar um usuário.",
        "parameters": [
          {
            "name": "userUpdate",
            "in": "body",
            "required": true,
            "description": "Informações do usuário.",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "api_key": {
              "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
              }
            }
          }
        ]
      }
    },
    "/users/{userDel}": {
      "delete": {
        "tags": [
          "User"
        ],
        "description": "Endpoint para excluir um usuário.",
        "parameters": [
          {
            "name": "userDel",
            "in": "body",
            "required": true,
            "description": "Informações do usuário.",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/DelUser"
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "api_key": {
              "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
              }
            }
          }
        ]
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "Jhon Doe"
        },
        "email": {
          "type": "string",
          "example": "jhondoe@teste.com"
        },
        "password": {
          "type": "string",
          "example": "123456"
        },
        "createAt": {
          "type": "string",
          "example": "2020-12-09 17:47:50"
        },
        "updateAt": {
          "type": "string",
          "example": "2020-12-09 17:47:50"
        }
      },
      "required": [
        "id",
        "name",
        "email",
        "password",
        "createAt",
        "updateAt"
      ],
      "xml": {
        "name": "User"
      }
    },
    "AddUser": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Jhon Doe"
        },
        "email": {
          "type": "string",
          "example": "jhondoe@teste.com"
        },
        "password": {
          "type": "string",
          "example": "12345"
        }
      },
      "required": [
        "name",
        "email",
        "password"
      ],
      "xml": {
        "name": "AddUser"
      }
    },
    "DelUser": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "id"
      ],
      "xml": {
        "name": "DelUser"
      }
    },
    "SignIn": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "adm@teste.com"
        },
        "password": {
          "type": "string",
          "example": "12345678"
        }
      },
      "required": [
        "email",
        "password"
      ],
      "xml": {
        "name": "SignIn"
      }
    }
  }
}